Refactor UcrParticipationContainer into ExplorerHeaderContainer
